+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Change log for TextGenerator
 - Add 'substring' filter to FilterFunction
 - Fix error when a filter doesn't exist
 - Fix set function when the variable name contains another existing variable name
+- Return an error in date filter is used with a wrong format
 
 ## 0.3.0 - September 05, 2016